dry-run for rebasing tags

+++ rebase-on-tag.sh
@@ -2,6 +2,11 @@
 
 set -euo pipefail
 
+push_enabled=$1
+if [ "$push_enabled" = "--push" ]; then
+	shift
+fi
+
 latest_tag=$1
 
 if git ls-remote --exit-code --tags origin "refs/tags/${latest_tag}-lurtz" > /dev/null 2>&1; then
@@ -18,8 +23,14 @@ git fetch evcc # evcc is upstream
 git checkout master-lurtz
 git pull
 git rebase $latest_tag
-git push --force-with-lease
 git tag $latest_tag-lurtz
-git push origin $latest_tag-lurtz
+
+if [ "$push_enabled" = "--push" ]; then
+	echo "Pushing tag ${latest_tag}-lurtz to origin."
+	git push --force-with-lease
+	git push origin $latest_tag-lurtz
+else
+	echo "Push is disabled. Skipping push of tag ${latest_tag}-lurtz to origin."
+fi
 
 exit 0
